feat: export SimpleRecord which has the data methods of Resource
