+++ packages/rest-hooks/src/resource/SimpleRecord.ts
@@ -1,0 +1,82 @@
+import { AbstractInstanceType } from '~/types';
+
+const DefinedMembersKey = Symbol('Defined Members');
+type Filter<T, U> = T extends U ? T : never;
+interface SimpleResourceMembers<T extends typeof SimpleRecord> {
+  [DefinedMembersKey]: Filter<keyof AbstractInstanceType<T>, string>[];
+}
+
+/** Immutable record that keeps track of which members are defined vs defaults. */
+export default abstract class SimpleRecord {
+  /** Factory method. Takes an object of properties to assign. */
+  static fromJS<T extends typeof SimpleRecord>(
+    this: T,
+    props: Partial<AbstractInstanceType<T>>,
+  ) {
+    // we type guarded abstract case above, so ok to force typescript to allow constructor call
+    const instance = new (this as any)(props) as Readonly<
+      AbstractInstanceType<T>
+    >;
+
+    Object.defineProperty(instance, DefinedMembersKey, {
+      value: Object.keys(props),
+      writable: false,
+    });
+
+    Object.assign(instance, props);
+
+    // to trick normalizr into thinking we're Immutable.js does it doesn't copy
+    Object.defineProperty(instance, '__ownerID', {
+      value: 1337,
+      writable: false,
+    });
+    return instance;
+  }
+
+  /** Creates new instance copying over defined values of arguments */
+  static merge<T extends typeof SimpleRecord>(
+    this: T,
+    first: AbstractInstanceType<T>,
+    second: AbstractInstanceType<T>,
+  ) {
+    const props = Object.assign(
+      {},
+      this.toObjectDefined(first),
+      this.toObjectDefined(second),
+    );
+    return this.fromJS(props);
+  }
+
+  /** Whether key is non-default */
+  static hasDefined<T extends typeof SimpleRecord>(
+    this: T,
+    instance: AbstractInstanceType<T>,
+    key: Filter<keyof AbstractInstanceType<T>, string>,
+  ) {
+    return ((instance as any) as SimpleResourceMembers<T>)[
+      DefinedMembersKey
+    ].includes(key);
+  }
+
+  /** Returns simple object with all the non-default members */
+  static toObjectDefined<T extends typeof SimpleRecord>(
+    this: T,
+    instance: AbstractInstanceType<T>,
+  ) {
+    const defined: Partial<AbstractInstanceType<T>> = {};
+    for (const member of ((instance as any) as SimpleResourceMembers<T>)[
+      DefinedMembersKey
+    ]) {
+      defined[member] = instance[member];
+    }
+    return defined;
+  }
+
+  /** Returns array of all keys that have values defined in instance */
+  static keysDefined<T extends typeof SimpleRecord>(
+    this: T,
+    instance: AbstractInstanceType<T>,
+  ) {
+    return ((instance as any) as SimpleResourceMembers<T>)[DefinedMembersKey];
+  }
+}
