+++ website/sidebars.json
@@ -66,7 +66,7 @@
       {
         "type": "subcategory",
         "label": "Interface Definitions",
-        "ids": ["api/resource", "api/FetchShape"]
+        "ids": ["api/resource", "api/SimpleRecord", "api/FetchShape"]
       },
       {
         "type": "subcategory",
